chore: Clean code
