+++ SwissTransferCore/TargetAssembly.swift
@@ -63,7 +63,7 @@ open class TargetAssembly {
 
                 #if DEBUG
                 return SwissTransferInjection(
-                    environment: STCore.ApiEnvironment.Prod(),
+                    environment: STCore.ApiEnvironment.Preprod(),
                     userAgent: UserAgentBuilder().userAgent,
                     databaseRootDirectory: realmRootDirectory
                 )
