+++ SwissTransferFeatures/TransferDetailsView/EmptyStates/VirusDetectedView.swift
@@ -31,8 +31,9 @@ struct VirusDetectedView: View {
             style: .emptyState
         )
         .padding(value: .medium)
+        .stNavigationBarStyle()
         .toolbar {
-            ToolbarItem(placement: .cancellationAction) {
+            ToolbarItem(placement: .topBarTrailing) {
                 Button(action: dismiss.callAsFunction) {
                     Image(systemName: "xmark")
                 }
