+++ SwissTransferFeatures/TransferDetailsView/TransferDetailsRootView.swift
@@ -90,6 +90,5 @@ public struct TransferDetailsRootView: View {
                 VirusDetectedView()
             }
         }
-        .stNavigationBarStyle()
     }
 }
